feat(actions): give the tree -> .md direction an executor
The .md -> tree direction has had a permanent executor since the file
watcher existed. The reverse had none: when ops arrived by sync, or a
render-affecting fix landed, a page's .md stayed wrong until somebody
happened to open it. A real 2,574-page workspace carried 704 stale pages,
702 of them safe to re-project with nothing removed.

The sweep writes only pages whose re-projection removes nothing from
disk. A page that would remove content is withheld for
`outl doctor --repair`, which backs every file up and applies volume
ceilings: a background pass that deletes is one that needs an undo, and
the command with the undo already exists. That gate also makes a torn op
log safe without the sweep knowing anything about op-log health, since a
truncated replay renders less than disk holds.

An undownloaded iCloud file reads as NotFound, which is indistinguishable
from "page deleted", so both the survey and the writer now ask
guard_absent_markdown — the same function, so a read-only listing cannot
promise a repair the writing pass refuses.
